DevKnow-backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
python-dotenv==1.2.2
django-cors-headers==4.9.0
openai==2.32.0
ruff==0.11.13